refactor: remove websocket package, rename leaderboardscoring -> leaderboardscoringapp
