--- leaderboardscoring/delivery/consumer/mapper.go
@@ -1,35 +1,0 @@
-package consumer
-
-import (
-	"fmt"
-	"github.com/gocasters/rankr/leaderboardscoring/service/leaderboardscoring"
-	"github.com/gocasters/rankr/protobuf/golang/eventpb"
-)
-
-func ProtobufToEventRequest(eventPB *eventpb.Event) (leaderboardscoring.EventRequest, error) {
-	if eventPB == nil {
-		return leaderboardscoring.EventRequest{}, fmt.Errorf("nil event")
-	}
-
-	payload := eventPB.GetContributionRegisteredPayload()
-	if payload == nil {
-		return leaderboardscoring.EventRequest{},
-			fmt.Errorf("event with ID %s has a missing or invalid payload", eventPB.Id)
-	}
-
-	ts := eventPB.GetTime()
-	if ts == nil {
-		return leaderboardscoring.EventRequest{}, fmt.Errorf("event with ID %s has a missing timestamp", eventPB.Id)
-	}
-	contributionEvent := leaderboardscoring.EventRequest{
-		ID:              eventPB.Id,
-		UserID:          payload.UserId,
-		ProjectID:       payload.ProjectId,
-		Type:            payload.EventType,
-		ScoreValue:      int(payload.ScoreValue),
-		SourceReference: payload.SourceReference,
-		Timestamp:       ts.AsTime().UTC(),
-	}
-
-	return contributionEvent, nil
-}
