--- leaderboardscoring/repository/dbconfig.local.yml
@@ -1,10 +1,0 @@
-postgres_db:
-  driver: postgres
-  host: localhost
-  port: 5432
-  user: leaderboardscoring_admin
-  password: password123
-  db_name: leaderboardscoring_db
-  ssl_mode: disable
-
-path_of_migration: ./leaderboardscoring/repository/migrations
