--- leaderboardscoring/service/leaderboardscoring/param.go
@@ -1,19 +1,0 @@
-package leaderboardscoring
-
-import "time"
-
-type EventRequest struct {
-	ID              string
-	UserID          string
-	ProjectID       string
-	Type            string
-	ScoreValue      int
-	SourceReference string
-	Timestamp       time.Time
-}
-
-type GetLeaderboardRequest struct {
-}
-
-type GetLeaderboardResponse struct {
-}
