+++ leaderboardscoringapp/service/leaderboardscoring/service.go
@@ -44,17 +44,19 @@ func (s Service) ProcessScoreEvent(ctx context.Context, req EventRequest) error 
 
 	contributionEvent := ContributionEvent{
 		ID:              req.ID,
-		UserID:          req.UserID,
-		ProjectID:       req.ProjectID,
-		Type:            ContributionType(req.Type),
-		ScoreValue:      req.ScoreValue,
+		Type:            ContributionType(""),
+		EventName:       req.EventName,
+		RepositoryID:    req.RepositoryID,
+		RepositoryName:  req.RepositoryName,
+		ContributorID:   req.ContributorID,
+		ScoreValue:      0,
 		SourceReference: req.SourceReference,
 		Timestamp:       req.Timestamp.UTC(),
 	}
 
-	var keys = s.keys(contributionEvent.ProjectID)
+	var keys = s.keys(contributionEvent.RepositoryName)
 
-	if err := s.repo.UpdateScores(ctx, keys, contributionEvent.ScoreValue, contributionEvent.UserID); err != nil {
+	if err := s.repo.UpdateScores(ctx, keys, contributionEvent.ScoreValue, contributionEvent.ContributorID); err != nil {
 		s.logger.Error(ErrFailedToUpdateScores.Error(), slog.String("error", err.Error()))
 		return err
 	}
