--- log/leaderboardscoring/service.log
@@ -1,13 +1,0 @@
-{"time":"2025-08-17T16:15:26.3702211+03:30","level":"INFO","msg":"start service"}
-{"time":"2025-08-17T16:18:29.0646301+03:30","level":"INFO","msg":"Starting leaderboardscoring Service..."}
-{"time":"2025-08-17T16:18:29.0721329+03:30","level":"ERROR","msg":"fatal error occurred","reason":"failed to connect to database","error":"failed to create pgx pool: MaxSize must be >= 1"}
-{"time":"2025-08-17T16:20:23.9900192+03:30","level":"INFO","msg":"Starting leaderboardscoring Service..."}
-{"time":"2025-08-17T16:20:23.9930608+03:30","level":"ERROR","msg":"fatal error occurred","reason":"failed to connect to database","error":"failed to create pgx pool: MaxSize must be >= 1"}
-{"time":"2025-08-17T16:20:23.998662+03:30","level":"INFO","msg":"Leaderboard-scoring service started"}
-{"time":"2025-08-17T16:20:51.0990621+03:30","level":"INFO","msg":"Starting leaderboardscoring Service..."}
-{"time":"2025-08-17T16:20:51.1036424+03:30","level":"ERROR","msg":"fatal error occurred","reason":"failed to connect to database","error":"failed to create pgx pool: MaxSize must be >= 1"}
-{"time":"2025-08-17T16:41:27.4341996+03:30","level":"INFO","msg":"Running migrations up..."}
-{"time":"2025-08-17T17:46:36.3877093+03:30","level":"INFO","msg":"Starting leaderboardscoring Service..."}
-{"time":"2025-08-17T17:46:36.3929652+03:30","level":"ERROR","msg":"fatal error occurred","reason":"failed to connect to database","error":"failed to create pgx pool: MaxSize must be >= 1"}
-{"time":"2025-08-17T18:08:19.7458581+03:30","level":"INFO","msg":"Running migrations up..."}
-{"time":"2025-08-17T18:08:38.8011397+03:30","level":"ERROR","msg":"invalid flags: --migrate-up and --migrate-down cannot be used together"}
